+++ scripts/update-test-cert.js
@@ -3,7 +3,7 @@ const https = require('https');
 const path = require('path');
 
 const SOURCE_HOST = 'httpbin.org';
-const TARGET_PATH = path.join(__dirname, '../test/app-template/www/certificates/httpbin.org.cer');
+const TARGET_PATH = path.join(__dirname, '../test/e2e-app-template/www/certificates/httpbin.org.cer');
 
 const getCert = hostname => new Promise((resolve, reject) => {
   const options = {